# Enable svt_av1_apply_temporal_filter_planewise_avx2 kernel