--- a/Source/Lib/Encoder/Codec/aom_dsp_rtcd.h
+++ b/Source/Lib/Encoder/Codec/aom_dsp_rtcd.h
@@ -1201,6 +1201,13 @@ extern "C" {
         uint64_t *mean_of_squared8x8_blocks);
     uint32_t sad_16bit_kernel_avx2(uint16_t *src, uint32_t src_stride, uint16_t *ref,
         uint32_t ref_stride, uint32_t height, uint32_t width);
+    void svt_av1_apply_temporal_filter_planewise_avx2(
+        const uint8_t *y_src, int y_src_stride, const uint8_t *y_pre, int y_pre_stride,
+        const uint8_t *u_src, const uint8_t *v_src, int uv_src_stride, const uint8_t *u_pre,
+        const uint8_t *v_pre, int uv_pre_stride, unsigned int block_width, unsigned int block_height,
+        int ss_x, int ss_y, const double *noise_levels, const int decay_control, uint32_t *y_accum,
+        uint16_t *y_count, uint32_t *u_accum, uint16_t *u_count, uint32_t *v_accum, uint16_t *v_count);
+
 #if RESTRUCTURE_SAD
 #endif
 #endif