Speedup compared to C implementation ~3x

For cmd line below:
```
-i 1920x1080_60fps_420_8bit.yuv -enc-mode [mode] -w 1920 -h 1080 -bit-depth 8 -fps 60 -b out.ivf
```

Encoding speedup:
  | Changes[fps] | master[fps]
-- | -- | --
M8 | 18.094 | 16.386
M7 | 11.487 | 11.131
M6 | 10.304 | 10.057
M5 | 5.546 | 5.491
M4 | 3.756 | 3.721
M3 | 2.456 | 2.418
M2 | 0.322 | 0.322
M1 | 0.052 | 0.052
M0 | 0.037 | 0.037