@spawlows 2020-05-08
Hi @hguermaz 

In PR #1120 

` SET_AVX2(svt_av1_apply_temporal_filter_planewise,
             svt_av1_apply_temporal_filter_planewise_c,
             svt_av1_apply_temporal_filter_planewise_c);`

you did not set kernel to svt_av1_apply_temporal_filter_planewise_avx2. It was by a mistake, or exist some reason?