@hguermaz 2020-05-11
Hi @spawlows ,

The files have been ported from cs2 to master, and cs2 has the problem.
The problem was introduced during cs2 clean up @ Revision: 86f93ba69d52c34874e1edc5bf787670722d8bbe, but we recently found there is no need to use C and that was a bug.