@OmranAbdelkafi 2020-05-16
Tested obj-1-fast for both 8 and 10 bit 


  | Speed Deviation | BD-Rate Deviation
-- | -- | --
M0 - 8bit | -0.38% | 0.00%
M4 - 8bit | -0.01% | 0.00%
M8 - 8bit | 8.74% | 0.00%
M0 - 10bit | 0.04% | 0.00%
M4 - 10bit | -1.17% | 0.00%
M8 - 10bit | 0.55% | 0.00%